requirements.txt
google-generativeai>=0.8.0
wechatpy>=1.8.18
requests>=2.31.0
python-dotenv>=1.0.0
markdown>=3.5.0
flask>=3.0.0
flask-cors>=4.0.0
zhipuai>=2.1.0
